+++ lib/image_processing/mini_magick.rb
@@ -21,15 +21,15 @@ module ImageProcessing
       IMAGE_CLASS = ::MiniMagick::Tool
 
       def resize_to_limit(magick, width, height)
-        magick.thumbnail "#{width}x#{height}>"
+        magick.resize "#{width}x#{height}>"
       end
 
       def resize_to_fit(magick, width, height)
-        magick.thumbnail "#{width}x#{height}"
+        magick.resize "#{width}x#{height}"
       end
 
       def resize_to_fill(magick, width, height, gravity: "Center")
-        magick.thumbnail "#{width}x#{height}^"
+        magick.resize "#{width}x#{height}^"
         magick.gravity gravity
         magick.background "rgba(255,255,255,0.0)" # transparent
         magick.extent "#{width}x#{height}"
@@ -38,7 +38,7 @@ module ImageProcessing
       def resize_and_pad(magick, width, height, background: :transparent, gravity: "Center")
         background = "rgba(255,255,255,0.0)" if background.to_s == "transparent"
 
-        magick.thumbnail "#{width}x#{height}"
+        magick.resize "#{width}x#{height}"
         magick.background background
         magick.gravity gravity
         magick.extent "#{width}x#{height}"